nous-py/pyproject.toml
[project]
name = "nous-ai"
version = "0.1.0"
description = "AI operations for Nous notebook application"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "openai>=1.0",
    "anthropic>=0.40",
    "boto3>=1.35",
    "tavily-python>=0.5",
    "trafilatura>=1.12",
    "markitdown>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
ollama = [
    "ollama>=0.4",
]
browser = [
    "browser-use>=0.2",
]
transcription = [
    "faster-whisper>=1.0",
]
audio = [
    "elevenlabs>=1.0",
    "pydub>=0.25",
]
kokoro = [
    "kokoro>=0.9",
    "soundfile>=0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
